move code and assets into source folder
