+++ tkinter-designer/core.py
