+++ tkinter-designer/images/iconbitmap.gif
